requirements.txt
flask==2.0.1
werkzeug==2.0.1
jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
pymongo==4.3.3
python-dotenv==0.19.1
flask-swagger-ui==3.36.0
